Rename loans layouts
