+++ lib/src/features/loans/presentation/loans_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-import 'desktop_layout.dart';
-import 'mobile_layout.dart';
+import 'loans_desktop_layout.dart';
+import 'loans_mobile_layout.dart';
 
 class LoansPage extends StatelessWidget {
   const LoansPage({super.key});
