+++ lib/src/features/loans/presentation/open_loan_page.dart
@@ -9,7 +9,7 @@ import 'package:librarian_app/src/features/loans/presentation/loan_details.dart'
 import 'package:librarian_app/src/features/borrowers/presentation/borrowers_list_view.dart';
 import 'package:provider/provider.dart';
 
-import 'mobile_layout.dart';
+import 'loans_mobile_layout.dart';
 
 class OpenLoanPage extends StatefulWidget {
   const OpenLoanPage({super.key});
